Split store logic test
