+++ packages/services/f-statistics/src/store/logic/_tests/shouldPublishQueuedLogs.test.js
@@ -1,6 +1,4 @@
-import isQueueLengthExceeded from '../isQueueLengthExceeded';
 import shouldPublishQueuedLogs from '../shouldPublishQueuedLogs';
-// import startBatchPublishTimer from '../startBatchPublishTimer;';
 import publishQueue from '../publishQueue';
 import * as config from '../../../config';
 import { log } from '../../../tests/helpers/setup';
@@ -14,44 +12,10 @@ jest.mock('../../../config', () => ({
 
 jest.mock('../publishQueue', () => jest.fn());
 
-describe('Statistics Utils ::', () => {
+describe('Statistics Store Logic ::', () => {
     afterEach(() => {
         jest.clearAllMocks();
     });
-
-    describe('isQueueLengthExceeded ::', () => {
-        it('should return false if queue is less than config.BATCH_QUEUE_SIZE', () => {
-            // Arrange
-            const queue = Array(3).fill(log);
-
-            // Act
-            const result = isQueueLengthExceeded(queue);
-
-            // Assert
-            expect(result).toBeFalsy();
-        });
-        it('should return true if queue is equal to config.BATCH_QUEUE_SIZE', () => {
-            // Arrange
-            const queue = Array(5).fill(log);
-
-            // Act
-            const result = isQueueLengthExceeded(queue);
-
-            // Assert
-            expect(result).toBeTruthy();
-        });
-        it('should return true if queue is greater than config.BATCH_QUEUE_SIZE', () => {
-            // Arrange
-            const queue = Array(6).fill(log);
-
-            // Act
-            const result = isQueueLengthExceeded(queue);
-
-            // Assert
-            expect(result).toBeTruthy();
-        });
-    });
-
 
     describe('shouldPublishQueuedLogs ::', () => {
         it('should call publishQueue() when batch publishing is not enabled config.IS_BATCH_PUBLISHING_ENABLED', () => {
